+++ src/avian/lzma.h
@@ -24,14 +24,14 @@ namespace vm {
 uint8_t* decodeLZMA(System* s,
                     avian::util::Alloc* a,
                     uint8_t* in,
-                    unsigned inSize,
-                    unsigned* outSize);
+                    size_t inSize,
+                    size_t* outSize);
 
 uint8_t* encodeLZMA(System* s,
                     avian::util::Alloc* a,
                     uint8_t* in,
-                    unsigned inSize,
-                    unsigned* outSize);
+                    size_t inSize,
+                    size_t* outSize);
 
 }  // namespace vm
